Fixed previously created Unit Tests
-CreateAppointmentService.spec.ts, CreateUserService.spec.ts,
UpdateUserAvatar.spec.ts, AuthenticateUserService.spec.ts:
  -In some cases changed the used Service to Repository when it was not the core
thing being tested, since it would show up in the coverage report when it should
not. Example: When testing to ensure we can not create two appointments at the
same time, previously we used service to create both appointments, but the first
appointment shouldn't appear on the coverage report, so we use repository for it
instead.
  -Included await before expect() because the executed function is async and the
Unit Test was passing when it shouldn't due to that.